Ignore I/O errors
Unless the I/O-error is transient, the panic message can't be displayed either.
If what was being printed was a fatal error, the process is going to exit
anyway. maybe with a custom exit value. If it was a warning or something
lessimportant, it's probably not worth likely silently aborting
the the process over.
For people with more complex needs, this library is probably too simple anyway.

This is a breaking change.